search: always return an allocated array for JSON marshal
In Go JSON marshaling will render 'null' for an unallocated array and
'[]' for an allocated array. This is unexpected by some JS apps that
will fail to treat null like an empty array.

As a result of our X-Namespace change the first query in
searchCustomers can return zero results which will return the
unallocated array to the UI.