keep both hashed_account_number and sha256_account_number updated
We need to update hashed_account_number together with
sha256_account_number as currently it is used in accounts table unique
constrain to validate the uniqueness of the number for customer.

Next release will remove this column and create unique constrain using
sha256_account_number instead of hashed_account_number.